+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Simple frontend with [Streamlit](https://www.streamlit.io/).
+
 ## [0.1.1] - 2020-07-04
 
 ### Added